Created module for popup windows
